# api.yaml
swagger: '2.0'
info:
  title: API микросервиса авторизации
  version: 1.0.0
paths:
  /auth/v1/mobile-login:
    post:
      description: Сервис авторизации для мобильных устройств
      tags:
        - MobileAuthService
      operationId: mobileAuthService
      produces:
        - application/json
      x-start-version: 1
      x-access: NoCheck
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/MobileLoginRequest'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/createSessionRs'
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        401:
          description: Ошибка аутентификации
        500:
          description: Internal server error
  /auth/v1/short-code:
    post:
      description: Регистрация короткого кода мобильного устройства
      tags:
        - MobileAuthService
      operationId: mobileAuthShortCodeService
      produces:
        - application/json
      x-start-version: 1
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ShortCodeRequest'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/ShortCodeResponse'
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        401:
          description: Ошибка аутентификации
        500:
          description: Internal server error
...
definitions:
...
  MobileLoginRequest:
    type: object
    description: |
      Параметры запроса авторизации
    required:
      - authMethod
      - login
      - password
    properties:
      authMethod:
        type: string
        enum:
          - AUTH_METH_NAME_AND_PASSWORD
          - AUTH_METH_SHORT_CODE
        description: |
          Метод авторизации
      login:
        type: string
        description: |
          Логин или токен
      password:
        type: string
        description: |
          Пароль или пин
      identifierDevice:
        type: string
        description: |
          Идентификатор устройства
      platform:
        type: string
        enum:
          - ANDROID
          - IOS
        description: |
          Платформа
  ShortCodeRequest:
    type: object
    description: |
      Информация о коротком коде
    required:
      - shortCode
      - sessionId
      - identifierDevice
      - platform
    properties:
      shortCode:
        type: string
        description: |
          Код доступа
      sessionId:
        type: string
        description: |
          id-сессии
      identifierDevice:
        type: string
        description: |
          Идентификатор устройства
      platform:
        type: string
        enum:
          - ANDROID
          - IOS
        description: |
          Платформа
  ShortCodeResponse:
    type: object
    description: |
      Информация об установке короткого кода
    required:
      - token
    properties:
      token:
        type: string
        description: |
          Токен мобильного устройства на сервере
...
  errorData:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            fieldName:
              type: string
              description: Поле, к которому относится ошибка
            errorMessage:
              type: string
              description: Сообщение об ошибке
...
  createSessionRs:
    type: object
    properties:
      id:
        type: string
        description: идентификатор сессии
      user:
        type: object
        properties:
          id:
            type: integer
            format: int64
            description: Идентификатор учётной записи сотрудника
          login:
            type: string
            description: Логин сотрудника
          secondName:
            type: string
            description: Фамилия сотрудника
          firstName:
            type: string
            description: Имя сотрудника
          middleName:
            type: string
            description: Отчество сотрудника
          mobilePhone:
            type: string
            description: Номер мобильного телефона
          email:
            type: string
            description: Адрес электронной почты
          position:
            type: string
            description: Должность сотрудника
          office:
            type: string
            description: Название ВСП
          branch:
            type: string
            description: Название регионального филиала
          offices:
            description: Список филиалов назначенных пользователю
            type: array
            items:
              $ref: '#/definitions/orgUnitData'
          personnelNumber:
            type: string
            description: Табельный номер
          rights:
            type: array
            description: Назначенные права
            items:
              type: string
              description: Имя права
          changePassword:
            type: boolean
            description: Признак необходимости сменить пароль при входе
          groupCodes:
            description: Список кодов групп пользователя
            type: array
            items:
              type: string
...
  orgUnitData:
    type: object
    description: Организационная штатная структура
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи
      branch:
        type: string
        minLength: 1
        maxLength: 255
        description: Название регионального филиала (РФ); заполняется, если type = РФ.
      office:
        type: string
        minLength: 1
        maxLength: 255
        description: Номер внутреннего структурного подразделения (ВСП); заполняется, если type = ВСП.
...