# auth-service-app/src/main/java/ru/softlab/efr/services/auth/controllers/MobileAuthServiceApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ru.softlab.efr.services.auth.controllers;

import ru.softlab.efr.services.auth.exchange.model.CreateSessionRs;
import ru.softlab.efr.services.auth.exchange.model.ErrorData;
import ru.softlab.efr.services.auth.exchange.model.MobileLoginRequest;
import ru.softlab.efr.services.auth.exchange.model.ShortCodeRequest;
import ru.softlab.efr.services.auth.exchange.model.ShortCodeResponse;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.validation.BindingResult;
import org.springframework.data.web.PageableDefault;
import org.springframework.data.domain.*;

import java.util.*;

import org.springframework.validation.annotation.Validated;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * the MobileAuthService API
 */
public interface MobileAuthServiceApi {

    /**
     * Сервис авторизации для мобильных устройств 
     * @param body   
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/auth/v1/mobile-login",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<CreateSessionRs> mobileAuthService(@Valid @RequestBody MobileLoginRequest body) throws Exception;



    /**
     * Регистрация короткого кода мобильного устройства 
     * @param body   
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = "/auth/v1/short-code",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<ShortCodeResponse> mobileAuthShortCodeService(@Valid @RequestBody ShortCodeRequest body) throws Exception;


}